use float32 for weights

+++ sbayes/util.py
@@ -27,7 +27,7 @@ from scipy.sparse import csr_matrix
 from numba import jit, njit, float32, float64, int64, boolean, vectorize
 
 
-FLOAT_TYPE = np.float64
+FLOAT_TYPE = np.float32
 INT_TYPE = np.int64
 EPS = np.finfo(float).eps
 RNG = np.random.default_rng()
